Refactor on feedback from pglusacTT
